requirements.txt
httpx>=0.28.1
python-decouple>=3.8
